Revert "Revert "2014年7月21日""
This reverts commit 0641fd626669c3826ce59efbb3b219bb93bbcfe7.

+++ CFJDS/Common/DataCFSJ.cs
@@ -100,9 +100,9 @@ namespace CFJDS {
         /// <summary>
         /// 违法单价
         /// </summary>
-        private int _illegaUnit;
+        private double _illegaUnit;
 
-        public int IllegaUnit {
+        public double IllegaUnit {
             get { return _illegaUnit; }
             set { _illegaUnit = value; }
         }
